refactor chat component, separate the chat input area dna message area to solve dynamic height sync issue.
